+++ cpp_src/mime/charset.cpp
@@ -12,7 +12,7 @@ zcc_namespace_begin;
 
 std::string mail_parser::charset_convert(const char *from_charset, const char *data, int64_t size)
 {
-    return charset::convert_to_utf8(from_charset, data, size);
+    return charset::convert_to_utf8(from_charset, data, (int)size);
 }
 
 zcc_namespace_end;
